change objective and base
